fix(6.4): RUF005, RUF036, RUF046, RUF059
Yettita mayda zamonaviylashtirish.

**RUF046 (4 ta)** — `int(round(x))` da `int()` ortiqcha: `round()`
argumentsiz chaqirilganda allaqachon butun son qaytaradi. Xulq-atvor
o'zgarmadi, faqat ikki marta o'girish olib tashlandi.

**RUF005 (1 ta)** — `templates + (yangi,)` o'rniga `(*templates, yangi)`.
Yangi shakl vaqtinchalik ro'yxat yaratmaydi va niyat aniqroq.

**RUF036 (1 ta)** — `A | None | object` da `None` oxirida bo'lishi
kerak: `A | object | None`. Faqat yozilish tartibi.

**RUF059 (1 ta)** — testda `engine` ochilardi, lekin ishlatilmasdi.
`_engine` deb belgilandi. Diqqat: shu fayldagi 355 va 386-qatorlarda
xuddi shunday ochish bor, lekin ular `engine` ni ishlatadi — shuning
uchun faqat 444-qator o'zgardi.

To'rtala qoida ham ignore ro'yxatidan o'chirildi.

+++ backend/app/business_online/payload/dining.py
@@ -54,7 +54,7 @@ def dining_prepared_items(
             continue
         quantity = wanted[item_id]
         price = parse_price_amount(item.get("price"))
-        line_total = int(round(price * quantity))
+        line_total = round(price * quantity)
         prepared.append(
             {
                 "item_id": item_id,
